requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
pdfminer.six>=20221105
gtts>=2.5.0
Werkzeug>=3.0.0